Update whoami canister
