--- examples/motoko_examples/whoami/benchmarks.csv
@@ -1,3 +1,0 @@
-Description,Azle Wasm Instructions (function body only),Azle Wasm Instructions (function prelude and body),Motoko Wasm Instructions (function body only),Motoko Wasm Instructions (function prelude and body),Rust Wasm Instructions (function body only),Rust Wasm Instructions (function prelude and body),Azle/Motoko Change Multiplier (function body only),Azle/Motoko Change Multiplier (function prelude and body),Azle/Rust Change Multiplier (function body only),Azle/Rust Change Multiplier (function prelude and body),Motoko/Azle Change Multiplier (function body only),Motoko/Azle Change Multiplier (function prelude and body),Motoko/Rust Change Multiplier (function body only),Motoko/Rust Change Multiplier (function prelude and body),Rust/Azle Change Multiplier (function body only),Rust/Azle Change Multiplier (function prelude and body),Rust/Motoko Change Multiplier (function body only),Rust/Motoko Change Multiplier (function prelude and body)
-id,50584105,50247985,430,4334,400,39861,117637,11594,126460,1261,-117637,-11594,1,-9,-126460,-1261,-1,9
-whoami,19374040,19912137,217,1465,1277,20928,89281,13592,15172,951,-89281,-13592,-6,-14,-15172,-951,6,14
