+++ examples/motoko_examples/whoami/dfx.json
@@ -1,31 +1,14 @@
 {
     "canisters": {
-        "azle": {
+        "whoami": {
             "type": "custom",
-            "build": "npx azle azle",
-            "root": "canisters/azle",
-            "ts": "canisters/azle/index.ts",
-            "candid": "canisters/azle/index.did",
-            "wasm": ".azle/azle/azle.wasm.gz",
+            "build": "npx azle whoami",
+            "root": "src",
+            "ts": "src/index.ts",
+            "candid": "src/index.did",
+            "wasm": ".azle/whoami/whoami.wasm.gz",
             "declarations": {
-                "output": "dfx_generated/azle",
-                "node_compatibility": true
-            }
-        },
-        "motoko": {
-            "type": "motoko",
-            "main": "canisters/motoko/Main.mo",
-            "declarations": {
-                "output": "dfx_generated/motoko",
-                "node_compatibility": true
-            }
-        },
-        "rust": {
-            "type": "rust",
-            "package": "rust",
-            "candid": "canisters/rust/src/lib.did",
-            "declarations": {
-                "output": "dfx_generated/rust",
+                "output": "./dfx_generated/whoami",
                 "node_compatibility": true
             }
         }
