+++ examples/motoko_examples/whoami/package.json
@@ -8,9 +8,9 @@
         "azle": "0.17.1"
     },
     "devDependencies": {
-        "@dfinity/agent": "0.11.1",
-        "@dfinity/identity": "0.11.1",
-        "ts-node": "10.7.0",
-        "typescript": "4.6.3"
+        "@dfinity/agent": "^0.19.2",
+        "@dfinity/identity": "^0.19.2",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.2.2"
     }
 }
