+++ examples/motoko_examples/whoami/src/index.did
