+++ examples/motoko_examples/whoami/test/test.ts
@@ -1,5 +1,5 @@
 import { runTests } from 'azle/test';
-import { createActor } from '../dfx_generated/azle';
+import { createActor } from './dfx_generated/whoami';
 import { callingIdentity, canisterId, getTests } from './tests';
 
 const whoami_canister = createActor(canisterId, {
